requirements.txt
# CivicAid Python Dependencies

# Core Web Framework
Flask>=3.0.0
flask-cors>=4.0.0
Werkzeug>=3.0.0
gunicorn>=21.0.0

# API & Environment
requests>=2.31.0
python-dotenv>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
google-generativeai>=0.3.0
